Convolution : change the default convolution
