+++ benchmarks/CMakeLists.txt
@@ -7,7 +7,7 @@ if (${USE_GPU})
     endif()
 endif()
 add_subdirectory(DNN/layers/convolution/direct/cpu)
-add_subdirectory(DNN/layers/convolution/direct/cpu_32channels)
+add_subdirectory(DNN/layers/convolution/direct/cpu_3channels)
 add_subdirectory(DNN/layers/convolution/direct/sparse)
 add_subdirectory(DNN/layers/convolution/direct/sparse_32channels)
 add_subdirectory(DNN/layers/convolution/direct/cpu_sparse_with_dense)
