--- benchmarks/DNN/layers/convolution/direct/cpu_32channels/clean.sh
@@ -1,1 +1,0 @@
-rm conv_layer_generator_32channels conv_layer_mkl conv_layer_mkldnn generated_conv_layer.o generated_conv_layer.o.h mkl_result.txt tiramisu_result.txt wrapper_conv_layer_32channels
