+++ benchmarks/DNN/layers/convolution/direct/cpu_3channels/clean.sh
@@ -1,0 +1,1 @@
+rm conv_layer_generator_3channels conv_layer_mkl conv_layer_mkldnn generated_conv_layer.o generated_conv_layer.o.h mkl_result.txt tiramisu_result.txt wrapper_conv_layer_3channels
