+++ benchmarks/DNN/layers/convolution/direct/cpu_3channels/compile_and_run_mkldnn.sh
