+++ benchmarks/DNN/layers/convolution/direct/cpu_3channels/conv_layer_generator_mkl.c
